chore(RHINENG-16141): Cleanup FinishedCreatePolicy
